eks_deploy/services/backend/requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.10.0
Flask-Cors==3.0.10
Flask-PyMongo==2.2.0
requests==2.20.1
gunicorn